feat: score as Double, usedColorsForGame in state, ObserveColorsUsedForGameUseCase, update tests
Agent-Logs-Url: https://github.com/trumpets/meeplebook/sessions/bb716fa9-4caf-4c28-a55d-7ccad1ae4df2

+++ app/src/test/java/app/meeplebook/feature/addplay/AddPlayTestFactory.kt
@@ -18,7 +18,7 @@ object AddPlayTestFactory {
     fun makePlayer(
         identity: PlayerIdentity = makeIdentity(),
         startPosition: Int = 1,
-        score: Int? = null,
+        score: Double? = null,
         isWinner: Boolean = false,
         color: String? = null
     ): PlayerEntryUi = PlayerEntryUi(
